Types types types!!!! God I love types

+++ src/main/java/gregtech/api/enums/ToolModes.java
@@ -7,8 +7,8 @@ public enum ToolModes {
 
     private final byte modeValue;
 
-    ToolModes(byte modeValue) {
-        this.modeValue = modeValue;
+    ToolModes(int modeValue) {
+        this.modeValue = (byte) modeValue;
     }
 
     public byte get() {
